Create business logic
- Add NotFoundException
- Add "findRatesByBrandIdAndProductId" query
- Add converters between layers with mapstruct
- Add domain model
- Connect layers
- Add and modify tests

+++ src/main/java/com/bcnc/prices/web/model/ProductPriceRs.java
@@ -12,11 +12,11 @@ import lombok.EqualsAndHashCode;
 @Schema(description = "Product price response")
 public class ProductPriceRs extends BaseDto {
   @Schema(description = "Id of the product", example = "1")
-  private int productId;
+  private long productId;
   @Schema(description = "Id of the chain", example = "1")
-  private int chainId;
+  private long chainId;
   @Schema(description = "Id of the rate", example = "1")
-  private int rateId;
+  private long rateId;
   @Schema(description = "Starting date of application", example = "2024-10-20T07:37:25.123Z")
   private LocalDateTime startDate;
   @Schema(description = "Ending date of application", example = "2024-10-20T07:37:25.123Z")
